Extractor de datos/requirements.txt
# Dependencias principales
requests>=2.28.1
python-dotenv>=0.20.0
colorama==0.4.6

# Para web scraping
beautifulsoup4>=4.12.2

# Para procesamiento de fechas
python-dateutil==2.8.2

geopy==2.4.0
# pandas==2.2.3 # Already installed
gspread==5.12.0
oauth2client==4.1.3
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2

# For Football Player Market Value API
# Using requests library for market value API calls